requirements.txt
anyio==3.2.1
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Django==3.1.7
django-compat==1.0.15
django-crontab==0.7.1
empyrical==0.5.5
h11==0.12.0
httpcore==0.13.6
idna==3.2
lxml==4.6.3
numpy==1.21.1
pandas==1.3.1
pandas-datareader==0.10.0
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rfc3986==1.5.0
scipy==1.7.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
urllib3==1.26.6
